% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterpdf-package.R
\docType{package}
\name{rasterpdf}
\alias{rasterpdf}
\alias{rasterpdf-package}
\title{rasterpdf: Plot Raster Graphics in PDFs.}
\description{
rasterpdf is an R package to plot raster graphics in PDF files. This can be
useful when one needs multipage documents, but the plots contain so many
individual elements that use of vector graphics (with \code{\link[grDevices:pdf]{grDevices::pdf()}})
results in inconveniently large file sizes. Internally, rasterpdf plots each
individual page as a PNG, and then combines them in one PDF file.
}
\section{Functions}{


\code{\link[=raster_pdf]{raster_pdf()}} Opens a raster graphics PDF device.

\code{\link[=dev.off]{dev.off()}} Closes a graphics device.
}

