% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_pdf.R
\name{raster_pdf}
\alias{raster_pdf}
\title{raster_pdf}
\usage{
raster_pdf(filename = "Rplots.pdf", width = 7, height = 7,
  units = c("in", "cm", "mm", "px"), res = 72L, ...)
}
\arguments{
\item{filename}{filename}

\item{width}{width}

\item{height}{height}

\item{units}{units}

\item{res}{res}

\item{...}{...}
}
\description{
lorem ipsum
}
